Do not Short Circuit/solo_bot_vk/requirements.txt
Flask
vk-api
Pillow
requests
gunicorn
